Yakovenko version 1
This is a draft of an agent based model that starts with perfect
equality and generates an unequal distribution of wealth by transferring
wealth between random pairs of agents.  As the file notes, it is
inspired by something I saw Yakovenko present at a conference.    The
next version should (1) set the seed on the random number generator (2)
move the location of the random number draw that designates agent j in
the pair to after the "if agent i has a non-zero balance" statement, so
that I can cut down on the number of draws (3) clean up the shape of the
code.  But I wanted to get something down on paper, so to speak.